(1446) Rename all lifeConditions to livingConditions
