+++ packages/api/server/services/shantytown/export/2_section_living_conditions/firePrevention.js
