+++ packages/frontend/webapp/src/js/app/mixins/livingConditionsMixin.js
@@ -1,4 +1,4 @@
-export const lifeConditionsMixin = {
+export const livingConditionsMixin = {
     props: {
         shantytown: {
             type: Object,
